Renesas (P): Downgrade to python2 again
repo nowadays warns if using python2,  but continues.  To avoid the
warning the the main development branch was updated to use python3
throughout.

However *if* python3 is used, repo requires it to be v 3.6 or later and
otherwise it fails.  The issue becomes that the older Ubuntu distro version
recommended by the documentation installs a too old python3 version by default.

Downgrading to python2 again is the easiest solution.  Note that is for Renesas
only, and only for Android P, so it will likely be better addressed for Android
10 and forward.